Merged after gating on a full census rather than the branch's own sample of
12 addresses: 824 catalogued bodies live in the 42 source files it touches.
The census found three regressions and none of them were the branch's - all
three are stale rows on master, corrected separately in b7ae8d7.

What it establishes, beyond the 68 -> 34 drop in failing units:

  - the PCH internal compiler error in the original brief does not exist;
  - C1060 was CAUSED by /Zm1000, which CMake appends to every MSVC below
    1310, not mitigated by it - /Zm reserves the arena up front and under
    wine 1000 does not fit alongside the mapped PCH;
  - src/OpenSMACX.def sitting inside if(MINGW) does NOT lose the export list
    under VC6, because __declspec(dllexport) makes it emit the decorated
    names natively; the .def is a mingw bridge and my brief was wrong;
  - VC6 will not route friendship through a privately inherited base, which
    is a real compiler defect and NOT worth fixing: making GraphicWin's
    inheritance public clears seven errors and drops the ratchet, because
    accessibility is an input to overload resolution.

That last one is the reusable lesson and the branch records it: building a
hierarchy both ways and diffing the objects is NOT sufficient evidence that a
header change is byte-neutral. Only recompiling the bodies is.

Nothing pushed.

# Conflicts:
#	docs/recovery/agent-findings.md

+++ src/sound.h
@@ -32,7 +32,7 @@ class DLLEXPORT Sound {
   Sound() { ; }
   ~Sound();
   int UNK1(int);
-  void fade(int a1);
+  void fade(unsigned long a1);
   int is_playing();
   int is_looping();
   int get_time();
